code reorg in meta crate
