+++ framework/meta/src/cmd/contract/sc_config.rs
@@ -1,18 +1,18 @@
-mod multi_contract_serde;
-mod oc_builder;
-mod oc_config;
-mod oc_global_config;
-mod oc_settings;
-mod oc_validate;
+mod contract_variant;
+mod contract_variant_builder;
+mod contract_variant_settings;
+mod contract_variant_validate;
+mod sc_config_model;
+mod sc_config_serde;
 mod wasm_build;
 mod wasm_clean;
 mod wasm_crate_gen;
 mod wasm_update;
 
-pub use multi_contract_serde::{
-    ContractVariantProfileSerde, ContractVariantSerde, MultiContractConfigSerde,
-    MultiContractGeneralSettingsSerde,
+pub use contract_variant::ContractVariant;
+pub use contract_variant_settings::{ContractVariantProfile, ContractVariantSettings};
+pub use sc_config_model::ScConfig;
+pub use sc_config_serde::{
+    ContractVariantProfileSerde, ContractVariantSerde, MultiContractGeneralSettingsSerde,
+    ScConfigSerde,
 };
-pub use oc_config::ContractVariant;
-pub use oc_global_config::ScConfig;
-pub use oc_settings::{ContractVariantProfile, ContractVariantSettings};
