+++ framework/meta/src/cmd/contract/sc_config/contract_variant.rs
