+++ framework/meta/src/cmd/contract/sc_config/contract_variant_builder.rs
@@ -5,11 +5,13 @@ use std::{
     path::{Path, PathBuf},
 };
 
+use crate::ei::parse_check_ei;
+
 use super::{
-    oc_global_config::SC_CONFIG_FILE_NAMES,
-    oc_settings::{parse_allocator, parse_check_ei, parse_stack_size},
+    contract_variant_settings::{parse_allocator, parse_stack_size},
+    sc_config_model::SC_CONFIG_FILE_NAMES,
     ContractVariant, ContractVariantProfile, ContractVariantSerde, ContractVariantSettings,
-    MultiContractConfigSerde, ScConfig,
+    ScConfig, ScConfigSerde,
 };
 
 /// Temporary structure, to help create instances of `ContractVariant`. Not publicly exposed.
@@ -252,7 +254,7 @@ impl ScConfig {
     /// Assembles an `ContractVariantConfig` from a raw config object that was loaded via Serde.
     ///
     /// In most cases the config will be loaded from a .toml file, use `load_from_file` for that.
-    pub fn load_from_config(config: &MultiContractConfigSerde, original_abi: &ContractAbi) -> Self {
+    pub fn load_from_config(config: &ScConfigSerde, original_abi: &ContractAbi) -> Self {
         let mut contract_builders: HashMap<String, ContractVariantBuilder> = config
             .contracts
             .iter()
@@ -297,7 +299,7 @@ impl ScConfig {
     pub fn load_from_file<P: AsRef<Path>>(path: P, original_abi: &ContractAbi) -> Option<Self> {
         match fs::read_to_string(path.as_ref()) {
             Ok(s) => {
-                let config_serde: MultiContractConfigSerde = toml::from_str(s.as_str())
+                let config_serde: ScConfigSerde = toml::from_str(s.as_str())
                     .unwrap_or_else(|error| panic!("error parsing multicontract.toml: {error}"));
                 Some(Self::load_from_config(&config_serde, original_abi))
             },
