+++ framework/meta/src/cmd/contract/sc_config/contract_variant_settings/stack_size.rs
