+++ framework/meta/src/ei/ei_version.rs
@@ -63,3 +63,18 @@ impl EIVersion {
         self.vm_hook_names().contains(&vm_hook_names)
     }
 }
+
+/// Parses an EIVersion, or returns None, if "ignore" was specifically stated.
+pub fn parse_check_ei(ei: &Option<String>) -> Option<EIVersion> {
+    if let Some(ei_name) = ei {
+        if ei_name == "ignore" {
+            None
+        } else {
+            let ei_version = EIVersion::from_name(ei_name)
+                .unwrap_or_else(|| panic!("invalid EI version: {ei_name}"));
+            Some(ei_version)
+        }
+    } else {
+        Some(EIVersion::default())
+    }
+}
