+++ framework/meta/tests/multi_contract_test.rs
@@ -1,7 +1,7 @@
 use multiversx_sc::abi::{ContractAbi, EndpointAbi};
-use multiversx_sc_meta::cmd::contract::sc_config::{MultiContractConfigSerde, ScConfig};
+use multiversx_sc_meta::cmd::contract::sc_config::{ScConfig, ScConfigSerde};
 
-fn get_serialized_toml() -> MultiContractConfigSerde {
+fn get_serialized_toml() -> ScConfigSerde {
     toml::from_str(
         r#"
         [settings]
